Fix pr #9: Fix issue #7: Implement Additional Database Models
