+++ users/tests/models/test_users.py
@@ -1,0 +1,2 @@
+import pytest
+from users.models import User
